logic/import_module.py
import config
from PIL import Image
from PyQt6.QtWidgets import QFileDialog


def import_image(layout, text, image_display):
    Image.MAX_IMAGE_PIXELS = config.MAX_IMAGE_PIXELS
    path, _ = QFileDialog.getOpenFileName(
        layout,
        text,
        "",
        "Images (*.png *.jpg *.jpeg *.bmp *.gif)"
    )
    if not path:
        return None, None

    imported_image = Image.open(path)
    image_display.set_image(imported_image)
    layout.button_gen_prov.setEnabled(True)
    
    return path, imported_image
