requirements.txt
numpy<1.25.0
pillow
PyQt6
scipy
tkinterdnd2
